# public/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Documentation",
    "version": "1.0.0",
    "description": "This is the OpenAPI specification for your project."
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/ab": {
      "get": {
        "operationId": "get-ab",
        "summary": " Get address book\r",
        "description": "Retrieve the user's address book",
        "tags": [
          "Ab"
        ],
        "parameters": [],
        "responses": {},
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "post": {
        "operationId": "post-ab",
        "summary": " Update address book\r",
        "description": "Update the user's address book with new peers and tags",
        "tags": [
          "Ab"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "tags",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "required": false
          },
          {
            "in": "query",
            "name": "peers",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "username": {
                    "type": "string"
                  },
                  "hostname": {
                    "type": "string"
                  },
                  "platform": {
                    "type": "string"
                  },
                  "alias": {
                    "type": "string"
                  },
                  "tags": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "hash": {
                    "type": "string"
                  }
                },
                "required": [
                  "id",
                  "username",
                  "hostname",
                  "platform",
                  "alias",
                  "tags",
                  "hash"
                ]
              }
            },
            "required": false
          },
          {
            "in": "query",
            "name": "tag_colors",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "tagColors",
            "schema": {
              "type": "object",
              "additionalProperties": {
                "type": "number"
              }
            },
            "required": false
          }
        ],
        "responses": {},
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/ab/peers": {
      "post": {
        "operationId": "post-ab-peers",
        "summary": " Retrieve a list of active peers for a user\r",
        "description": "Get paginated list of active peers with optional address book filtering",
        "tags": [
          "Ab"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "current",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {},
            "required": false
          },
          {
            "in": "query",
            "name": "ab",
            "schema": {},
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "username": {
                        "type": "string"
                      },
                      "hostname": {
                        "type": "string"
                      },
                      "platform": {
                        "type": "string"
                      },
                      "alias": {
                        "type": "string"
                      },
                      "tags": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "hash": {
                        "type": "string"
                      }
                    },
                    "required": [
                      "id",
                      "username",
                      "hostname",
                      "platform",
                      "alias",
                      "tags",
                      "hash"
                    ]
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/oidc/auth": {
      "post": {
        "operationId": "post-oidc-auth",
        "summary": " Initiate OIDC authentication\r",
        "description": "Start the OIDC authentication process for a device",
        "tags": [
          "Oidc"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "op",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "uuid",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "code": {
                      "type": "string"
                    },
                    "url": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "code",
                    "url"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/oidc/auth-query": {
      "get": {
        "operationId": "get-oidc-auth-query",
        "summary": " Query OIDC authentication status\r",
        "description": "Check the status of an OIDC authentication process",
        "tags": [
          "Oidc"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "id",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "uuid",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {},
                    "type": {},
                    "tfa_type": {},
                    "secret": {},
                    "user": {}
                  },
                  "required": [
                    "access_token",
                    "type",
                    "tfa_type",
                    "secret",
                    "user"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/oidc/query": {
      "post": {
        "operationId": "post-oidc-query",
        "summary": " Query OIDC authentication code\r",
        "description": "Check if an OIDC authentication code is ready",
        "tags": [
          "Oidc"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "schema": {
              "type": "string"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "access_token": {},
                    "type": {},
                    "tfa_type": {},
                    "secret": {},
                    "user": {}
                  },
                  "required": [
                    "access_token",
                    "type",
                    "tfa_type",
                    "secret",
                    "user"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  }
}